# tp-parametrage-front/src/app/services/models/academie-year-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface AcademieYearRequest {
  archive: boolean;
  description: string;
  id?: number;
  name: string;
}