# tp-parametrage-front/src/app/services/models/classe-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ClasseRequest {
  archive: boolean;
  description: string;
  id?: number;
  name: string;
}