# tp-parametrage-front/src/app/services/models/classe-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ClasseResponse {
  archive?: boolean;
  description?: string;
  id?: number;
  name?: string;
}