# tp-parametrage-front/src/app/services/models/program-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ProgramRequest {
  archive: boolean;
  description: string;
  id?: number;
  name: string;
}