# tp-parametrage-front/src/app/services/models/program-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ProgramResponse {
  archive?: boolean;
  description?: string;
  id?: number;
  name?: string;
}