# tp-parametrage-front/src/app/services/models/registration-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface RegistrationResponse {
  academieName?: string;
  academieYearId?: number;
  archive?: boolean;
  classeId?: number;
  classeName?: string;
  date?: string;
  description?: string;
  id?: number;
  programId?: number;
  programName?: string;
  registrationNu?: string;
  studentFistName?: string;
  studentId?: number;
  studentLastName?: string;
}