# tp-parametrage-front/src/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/students": {
      "get": {
        "tags": [
          "Students"
        ],
        "operationId": "getStudents",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StudentResponse"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Students"
        ],
        "operationId": "updateStudent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Students"
        ],
        "operationId": "saveStudent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentResponse"
                }
              }
            }
          }
        }
      }
    },
    "/registrations": {
      "get": {
        "tags": [
          "Registrations"
        ],
        "operationId": "getAllRegistrations",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RegistrationResponse"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Registrations"
        ],
        "operationId": "updateRegistration",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Registrations"
        ],
        "operationId": "addRegistration",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/programs": {
      "get": {
        "tags": [
          "Programs"
        ],
        "operationId": "getAllPrograms",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProgramResponse"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Programs"
        ],
        "operationId": "updateProgram",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProgramRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgramResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Programs"
        ],
        "operationId": "addProgram",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProgramRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgramResponse"
                }
              }
            }
          }
        }
      }
    },
    "/classes": {
      "get": {
        "tags": [
          "Classes"
        ],
        "operationId": "getAllClasses",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClasseResponse"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Classes"
        ],
        "operationId": "updateClasse",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClasseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClasseResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Classes"
        ],
        "operationId": "addClasse",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClasseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClasseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/academieYears": {
      "get": {
        "tags": [
          "AcademieYears"
        ],
        "operationId": "getAllAcademieYears",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AcademieYearResponse"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "AcademieYears"
        ],
        "operationId": "updateAcademieYear",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcademieYearRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcademieYearResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "AcademieYears"
        ],
        "operationId": "addKind",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcademieYearRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcademieYearResponse"
                }
              }
            }
          }
        }
      }
    },
    "/classes/{classeId}/assign-subject/{subjectId}": {
      "post": {
        "tags": [
          "Classes"
        ],
        "operationId": "assignSubjectToClasse",
        "parameters": [
          {
            "name": "classeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "subjectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClasseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/students/{id}": {
      "get": {
        "tags": [
          "Students"
        ],
        "operationId": "getStudentById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Students"
        ],
        "operationId": "deleteStudent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/registrations/{id}": {
      "get": {
        "tags": [
          "Registrations"
        ],
        "operationId": "getRegistrationById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Registrations"
        ],
        "operationId": "deleteRegistrationById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/programs/{id}": {
      "get": {
        "tags": [
          "Programs"
        ],
        "operationId": "getProgramById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgramResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Programs"
        ],
        "operationId": "deleteProgramById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/classes/{id}": {
      "get": {
        "tags": [
          "Classes"
        ],
        "operationId": "getClasseById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClasseResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Classes"
        ],
        "operationId": "deleteClasseById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/academieYears/{id}": {
      "get": {
        "tags": [
          "AcademieYears"
        ],
        "operationId": "getAcademieYearById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcademieYearResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "AcademieYears"
        ],
        "operationId": "deleteAcademieYearById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "StudentRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "emailPro": {
            "type": "string"
          },
          "emailPerso": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string",
            "pattern": "^\\+?[0-9]{10,15}$"
          },
          "address": {
            "type": "string"
          },
          "archive": {
            "type": "boolean"
          },
          "registrationNu": {
            "type": "string"
          }
        },
        "required": [
          "address",
          "archive",
          "emailPerso",
          "emailPro",
          "firstName",
          "lastName",
          "phoneNumber",
          "registrationNu"
        ]
      },
      "StudentResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "emailPro": {
            "type": "string"
          },
          "emailPerso": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "archive": {
            "type": "boolean"
          },
          "registrationNu": {
            "type": "string"
          }
        }
      },
      "RegistrationRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "description": {
            "type": "string"
          },
          "archive": {
            "type": "boolean"
          },
          "studentId": {
            "type": "integer",
            "format": "int64"
          },
          "classeId": {
            "type": "integer",
            "format": "int64"
          },
          "programId": {
            "type": "integer",
            "format": "int64"
          },
          "academieYearId": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "academieYearId",
          "archive",
          "classeId",
          "date",
          "description",
          "programId",
          "studentId"
        ]
      },
      "RegistrationResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "description": {
            "type": "string"
          },
          "archive": {
            "type": "boolean"
          },
          "studentId": {
            "type": "integer",
            "format": "int64"
          },
          "classeId": {
            "type": "integer",
            "format": "int64"
          },
          "programId": {
            "type": "integer",
            "format": "int64"
          },
          "academieYearId": {
            "type": "integer",
            "format": "int64"
          },
          "studentFistName": {
            "type": "string"
          },
          "studentLastName": {
            "type": "string"
          },
          "registrationNu": {
            "type": "string"
          },
          "classeName": {
            "type": "string"
          },
          "programName": {
            "type": "string"
          },
          "academieName": {
            "type": "string"
          }
        }
      },
      "ProgramRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "archive": {
            "type": "boolean"
          }
        },
        "required": [
          "archive",
          "description",
          "name"
        ]
      },
      "ProgramResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "archive": {
            "type": "boolean"
          }
        }
      },
      "ClasseRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "archive": {
            "type": "boolean"
          }
        },
        "required": [
          "archive",
          "description",
          "name"
        ]
      },
      "ClasseResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "archive": {
            "type": "boolean"
          }
        }
      },
      "AcademieYearRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "archive": {
            "type": "boolean"
          }
        },
        "required": [
          "archive",
          "description",
          "name"
        ]
      },
      "AcademieYearResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "archive": {
            "type": "boolean"
          }
        }
      }
    }
  }
}